More type mismatch fixing.

+++ maestro/include/MotorBoard.h
@@ -48,7 +48,7 @@ class MotorBoard{// : public RTT::TaskContext {
         void addMotor(HuboMotor* motor, int channel);
         void removeMotor(HuboMotor* motor);
         void removeMotor(int channel);
-        void setTicksPosition(vector<int> ticks);
+        void setTicksPosition(vector<long> ticks);
         HuboMotor* getMotorByChannel(int channel);
 
         // PROTOCOL MOTOR COMMANDS
